Move the JAVM run_instance lowering into JavmLocal
The InstanceCap+ImageCap -> ProgramSpec lowering half of
nub-arch-local (DATA_BASE, mem seed, pinned RoOverlay clamping,
endpoint Key lookup, reg layering — all gas/semantics-bearing, moved
verbatim) now lives next to the cap resolution in
javm::JavmLocal::invoke, tailing into the personality-agnostic
nub_arch_local::run_program. nub-arch-local ends javm-free
(deps: nub-kernel, nub-arch-x86-abi, javm-exec); javm gains javm-exec
for the interpreter register file. Conformance suite confirms
interp==recomp byte parity across the move.
