+++ rust/nub-arch-local/Cargo.toml
@@ -8,5 +8,4 @@ description = "Nub Arch implementation: in-process. Simulates the CPU/MMU substr
 [dependencies]
 nub-kernel = { path = "../nub-kernel" }
 nub-arch-x86-abi = { path = "../nub-arch-x86-abi" }
-javm-cap = { path = "../javm-cap" }
 javm-exec = { path = "../javm-exec" }
