+++ rust/nub-arch-local/src/lib.rs
@@ -2,22 +2,15 @@
 //! Rust data structures. Runs directly in the host process; no
 //! sandbox, no cross-compilation.
 //!
-//! Split in two layers:
-//!
-//! - [`run_program`] — personality-agnostic: takes a prepared
-//!   [`ProgramSpec`] (code, flat memory image, overlays, registers)
-//!   plus an [`EcallHandler`], wires it to
-//!   [`javm_exec::interp::Interpreter::run`], and produces an
-//!   [`InvocationResult`]. This is the in-process counterpart to
-//!   nub-arch-x86's JIT-driven `enter_frame` / `build_frame_runtime`.
-//! - [`run_instance`] — the JAVM wiring: lowers a published
-//!   [`javm_cap::cap::instance::InstanceCap`] + its referenced
-//!   [`javm_cap::cap::image::ImageCap`] into a [`ProgramSpec`].
-//!   (Temporary home — this half moves to the `javm` entrypoint crate
-//!   with the host personality.)
+//! Personality-agnostic: [`run_program`] takes a prepared
+//! [`ProgramSpec`] (code, flat memory image, overlays, registers)
+//! plus an [`EcallHandler`], wires it to
+//! [`javm_exec::interp::Interpreter::run`], and produces an
+//! [`InvocationResult`]. This is the in-process counterpart to
+//! nub-arch-x86's JIT-driven `enter_frame` / `build_frame_runtime`.
+//! The personality lowers its own object types into a `ProgramSpec`
+//! (JAVM: `javm::JavmLocal`'s `run_instance`).
 
-use javm_cap::cap::image::ImageCap;
-use javm_cap::cap::instance::InstanceCap;
 use javm_exec::{
     Access, CopyingMemory, EcallHandler, EcallKind, EcallResult, ExitReason, GasCounter, PAGE_SIZE,
     Regs, gas_const, interp::Interpreter, predecode::predecode_rv_with_mem_cycles,
@@ -76,8 +69,8 @@ pub struct RoOverlay {
 
 /// Personality-agnostic program description for [`run_program`]: what
 /// to execute, over what memory, starting from which register file.
-/// The personality (JAVM: [`run_instance`]) is responsible for
-/// lowering its own object types into this shape.
+/// The personality (JAVM: `javm::JavmLocal`'s `run_instance`) is
+/// responsible for lowering its own object types into this shape.
 pub struct ProgramSpec<'a> {
     /// The executable code region, mapped RO with PC = `code_base` +
     /// byte offset.
@@ -187,94 +180,6 @@ pub fn run_program(
         gas_remaining: gas.remaining(),
         scratchpad_head,
     }
-}
-
-/// Run an Instance through the PVM2 (RISC-V) interpreter by lowering
-/// the JAVM cap layout into a [`ProgramSpec`].
-///
-/// Endpoint dispatch: `endpoint_idx` selects
-/// `image.endpoints[endpoint_idx]`; the endpoint's `entry_pc` is used
-/// as the start PC. Caller-supplied `args` overlay φ[7..=10] on top
-/// of the endpoint's `initial_regs`. Memory is seeded from the
-/// Instance's `mem` DataCap (the whole RW extent), with pinned mappings
-/// re-laid read-only.
-pub fn run_instance(
-    instance: &InstanceCap,
-    image: &ImageCap,
-    endpoint_idx: u8,
-    args: [u64; 4],
-    initial_gas: u64,
-) -> InvocationResult {
-    let data_base = javm_cap::layout::DATA_BASE;
-    let data_extent = instance.mem.content_len();
-    let mut mem_image = vec![0u8; data_extent as usize];
-    if data_extent > 0 {
-        // Seed the whole extent from the Instance's memory image (the immutable
-        // backing — both initial and pinned content). No cache lookup needed.
-        instance.mem.copy_into(0, &mut mem_image);
-    }
-    // Pinned mappings become read-only re-lays (same bytes, from the
-    // seeded image) so a guest store faults, matching the recompiler's
-    // PinnedCapRo direct map.
-    let mut ro_overlays = Vec::new();
-    for m in image.mappings.iter() {
-        if m.path().is_empty() || !image.mapping_is_pinned(m.start as u32) {
-            continue;
-        }
-        let off = (m.start.saturating_sub(data_base as u64)) as usize;
-        let len = (m.size as usize).min(mem_image.len().saturating_sub(off));
-        if len > 0 {
-            ro_overlays.push(RoOverlay {
-                start: m.start as u32,
-                image_off: off,
-                len,
-            });
-        }
-    }
-
-    // V1 single-byte ABI: the endpoint selector is a single-byte Key into the
-    // sparse endpoint list.
-    let target = javm_cap::Key::from(endpoint_idx);
-    let (_, endpoint) = image
-        .endpoints
-        .iter()
-        .find(|(k, _)| *k == target)
-        .expect("endpoint key not defined");
-
-    let mut regs = Regs::new();
-    regs.pc = endpoint.entry_pc;
-    // Endpoint baseline first, then layer the InstanceCap's persisted
-    // regs on top (publish_instance writes them; subsequent invokes
-    // observe them). Args overlay φ[7..=10] last.
-    // Persisted file is the 13 host-mapped slots; x3/x4 (slots 13/14) start
-    // at 0 (Regs::new zeros them), matching the recompiler.
-    regs.gpr[..javm_cap::NUM_REGS].copy_from_slice(&endpoint.initial_regs);
-    for (i, v) in instance.regs.iter().enumerate() {
-        if *v != 0 {
-            regs.gpr[i] = *v;
-        }
-    }
-    for (i, v) in args.iter().enumerate() {
-        regs.gpr[7 + i] = *v;
-    }
-
-    // The executable code region, mapped RO at the fixed CODE_BASE
-    // (PC = CODE_BASE + byte_offset).
-    let (code_base, code_bytes) = image
-        .code_mapping()
-        .expect("image has no executable code mapping");
-
-    let spec = ProgramSpec {
-        code_base,
-        code: code_bytes,
-        data_base,
-        mem_image: &mem_image,
-        ro_overlays: &ro_overlays,
-        declared_mem_size: instance.mem_size(),
-        regs,
-    };
-    let mut handler = ExitingEcallHandler;
-    run_program(&spec, &mut handler, initial_gas)
 }
 
 fn page_round_up_u64(n: u64) -> u64 {
